pyproject.toml
[project]
name = "user-auth-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
